first-env-project/requirements.txt
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
distlib==0.3.4
elasticsearch==7.15.1
filelock==3.4.0
mysql-connector-python==8.0.27
platformdirs==2.4.0
protobuf==3.19.1
six==1.16.0
urllib3==1.26.7
virtualenv==20.10.0
